Remove requirement for component's ID to match that of template
The contents of the component will be appended to the relevant div,
rather than replacing it. This means that any ids, classes, styles
that were defined on the template element will not be overwritten,
and they will not need to be repeated in the component's top-level
element. Also means that the same component can be styled differently,
since it can be selected with the id corresponding to the component,
but have additional classes depending on where it is defined.

Also documented the steps required to add a page.